Exploring fix to use dynamic filter for aws instances inventory
Today aws_status_check.yaml has hardcoded the hostname with
the email of the user who deployed the environment in this form:
tag_instance_filter_three_tier_app_{{email_with_underscores}}
where {{email_with_underscores}} means underscores "_" instead
of "@" or "." for example foo-bar@example.com results in
foo_bar_example_com. This tag is created on aws inventory to
identify the instances created. This patch transform the email
with underscores and pass it as a variable to the playbook
aws_status_check.yaml